backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# CORS and middleware
starlette==0.27.0

# Computer Vision and ML
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.0
joblib==1.3.2

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Async support
aiofiles==23.2.1
aiohttp==3.9.1

# Image processing
imageio==2.31.6
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
click==8.1.7
typing-extensions==4.8.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Performance monitoring
psutil==5.9.6

# Logging
loguru==0.7.2

# Optional: TensorFlow for advanced ML (uncomment if needed)
# tensorflow==2.13.0
# tensorflow-cpu==2.13.0

# Optional: PyTorch for advanced ML (uncomment if needed)
# torch==2.1.0
# torchvision==0.16.0

# Optional: ONNX for model optimization (uncomment if needed)
# onnx==1.15.0
# onnxruntime==1.16.3